fix(browser): platform code cleanup
