+++ modules/@angular/compiler_cli/integrationtest/src/bootstrap.ts
@@ -1,8 +1,8 @@
 import {coreBootstrap, ReflectiveInjector} from '@angular/core';
-import {browserPlatform, BROWSER_APP_STATIC_PROVIDERS} from '@angular/platform-browser';
+import {browserPlatform, BROWSER_APP_PROVIDERS} from '@angular/platform-browser';
 import {BasicNgFactory} from './basic.ngfactory';
 import {Basic} from './basic';
 
 const appInjector =
-    ReflectiveInjector.resolveAndCreate(BROWSER_APP_STATIC_PROVIDERS, browserPlatform().injector);
+    ReflectiveInjector.resolveAndCreate(BROWSER_APP_PROVIDERS, browserPlatform().injector);
 coreBootstrap(BasicNgFactory, appInjector);
