+++ modules/@angular/compiler_cli/integrationtest/test/projection_spec.ts
@@ -1,12 +1,11 @@
 import {MainCompNgFactory} from '../src/projection.ngfactory';
 import {CompWithProjection} from '../src/projection';
 import {ReflectiveInjector, DebugElement, getDebugNode} from '@angular/core';
-import {browserPlatform, BROWSER_APP_STATIC_PROVIDERS, By} from '@angular/platform-browser';
+import {browserPlatform, BROWSER_APP_PROVIDERS, By} from '@angular/platform-browser';
 
 describe("content projection", () => {
   it("should support basic content projection", () => {
-    const appInjector = ReflectiveInjector.resolveAndCreate(BROWSER_APP_STATIC_PROVIDERS,
-                                                            browserPlatform().injector);
+    const appInjector = ReflectiveInjector.resolveAndCreate(BROWSER_APP_PROVIDERS, browserPlatform().injector);
     var mainComp = MainCompNgFactory.create(appInjector);
 
     var debugElement = <DebugElement>getDebugNode(mainComp.location.nativeElement);
