+++ modules/@angular/examples/core/ts/platform/platform.ts
@@ -1,5 +1,5 @@
 import {Component, createPlatform, coreLoadAndBootstrap, ReflectiveInjector} from '@angular/core';
-import {BROWSER_PLATFORM_PROVIDERS, BROWSER_APP_STATIC_PROVIDERS} from '@angular/platform-browser';
+import {BROWSER_PLATFORM_PROVIDERS, BROWSER_APP_PROVIDERS} from '@angular/platform-browser';
 
 var appProviders: any[] = [];
 
@@ -9,7 +9,6 @@ class MyApp {
 }
 
 var platform = createPlatform(ReflectiveInjector.resolveAndCreate(BROWSER_PLATFORM_PROVIDERS));
-var appInjector = ReflectiveInjector.resolveAndCreate([BROWSER_APP_STATIC_PROVIDERS, appProviders],
-                                                      platform.injector);
+var appInjector = ReflectiveInjector.resolveAndCreate([BROWSER_APP_PROVIDERS, appProviders], platform.injector);
 coreLoadAndBootstrap(MyApp, appInjector);
 // #enddocregion
