+++ modules/@angular/integration_test/public_api_spec.ts
@@ -370,12 +370,8 @@ var PLATFORM_BROWSER: string[] = [
   'bootstrap',
   'bootstrapApp',
   'bootstrapRender',
-  'bootstrapStatic',
-  'bootstrapStaticApp',
-  'bootstrapStaticRender',
-  'BROWSER_APP_COMMON_PROVIDERS',
   'BROWSER_APP_PROVIDERS',
-  'BROWSER_APP_STATIC_PROVIDERS',
+  'BROWSER_APP_COMPILER_PROVIDERS',
   'BROWSER_PLATFORM_PROVIDERS',
   'BROWSER_SANITIZATION_PROVIDERS',
   'browserPlatform',
@@ -408,18 +404,13 @@ var PLATFORM_BROWSER: string[] = [
   'UrlChangeEvent:dart',
   'UrlChangeListener:dart',
   'WebWorkerInstance',
-  'WORKER_APP_APPLICATION_COMMON_PROVIDERS',
   'WORKER_APP_APPLICATION_PROVIDERS',
   'WORKER_APP_LOCATION_PROVIDERS',
   'WORKER_APP_PLATFORM_PROVIDERS',
-  'WORKER_APP_STATIC_APPLICATION_PROVIDERS',
-  'WORKER_RENDER_APPLICATION_COMMON_PROVIDERS',
-  'WORKER_RENDER_APPLICATION_PROVIDERS',
   'WORKER_RENDER_APPLICATION_PROVIDERS',
   'WORKER_RENDER_LOCATION_PROVIDERS',
   'WORKER_RENDER_PLATFORM_PROVIDERS',
   'WORKER_RENDER_STARTABLE_MESSAGING_SERVICE',
-  'WORKER_RENDER_STATIC_APPLICATION_PROVIDERS',
   'WORKER_SCRIPT',
   'workerAppPlatform',
   'workerRenderPlatform'
