+++ modules/@angular/platform-browser/src/dom/debug/ng_probe.ts
@@ -12,7 +12,7 @@ import {getDOM} from '../dom_adapter';
 import {DomRootRenderer} from '../dom_renderer';
 
 
-const CORE_TOKENS = /*@ts2dart_const*/ {'ApplicationRef': ApplicationRef, 'NgZone': NgZone};
+const CORE_TOKENS = {'ApplicationRef': ApplicationRef, 'NgZone': NgZone};
 
 const INSPECT_GLOBAL_NAME = 'ng.probe';
 const CORE_TOKENS_GLOBAL_NAME = 'ng.coreTokens';
@@ -42,16 +42,16 @@ function _createRootRenderer(rootRenderer) {
 /**
  * Providers which support debugging Angular applications (e.g. via `ng.probe`).
  */
-export const ELEMENT_PROBE_PROVIDERS: any[] = /*@ts2dart_const*/[
-  /*@ts2dart_Provider*/ {
+export const ELEMENT_PROBE_PROVIDERS: any[] = [
+  {
     provide: RootRenderer,
     useFactory: _createConditionalRootRenderer,
     deps: [DomRootRenderer]
   }
 ];
 
-export const ELEMENT_PROBE_PROVIDERS_PROD_MODE: any[] = /*@ts2dart_const*/[
-  /*@ts2dart_Provider*/ {
+export const ELEMENT_PROBE_PROVIDERS_PROD_MODE: any[] = [
+  {
     provide: RootRenderer,
     useFactory: _createRootRenderer,
     deps: [DomRootRenderer]
