+++ modules/@angular/platform-browser/src/web_workers/shared/api.ts
@@ -1,3 +1,3 @@
 import {OpaqueToken} from '@angular/core';
 
-export const ON_WEB_WORKER = /*@ts2dart_const*/ new OpaqueToken('WebWorker.onWebWorker');
+export const ON_WEB_WORKER = new OpaqueToken('WebWorker.onWebWorker');
