+++ modules/@angular/platform-browser/src/web_workers/shared/serializer.ts
@@ -8,7 +8,7 @@ import {LocationType} from './serialized_types';
 
 // PRIMITIVE is any type that does not need to be serialized (string, number, boolean)
 // We set it to String so that it is considered a Type.
-export const PRIMITIVE: Type = /*@ts2dart_const*/ String;
+export const PRIMITIVE: Type = String;
 
 @Injectable()
 export class Serializer {
