+++ modules/@angular/platform-browser/src/web_workers/ui/location_providers.ts
@@ -6,10 +6,10 @@ import {APP_INITIALIZER, Injector, NgZone} from '@angular/core';
  * A list of {@link Provider}s. To use the router in a Worker enabled application you must
  * include these providers when setting up the render thread.
  */
-export const WORKER_RENDER_LOCATION_PROVIDERS = /*@ts2dart_const*/[
+export const WORKER_RENDER_LOCATION_PROVIDERS = [
   MessageBasedPlatformLocation,
   BrowserPlatformLocation,
-  /* @ts2dart_Provider */ {provide: APP_INITIALIZER, useFactory: initUiLocation, multi: true, deps: [Injector]}
+  {provide: APP_INITIALIZER, useFactory: initUiLocation, multi: true, deps: [Injector]}
 ];
 
 function initUiLocation(injector: Injector): () => void {
