+++ modules/@angular/platform-browser/src/web_workers/worker/location_providers.ts
@@ -6,8 +6,8 @@ import {WebWorkerPlatformLocation} from './platform_location';
  * Those providers should be added when the router is used in a worker context in addition to the
  * {@link ROUTER_PROVIDERS} and after them.
  */
-export const WORKER_APP_LOCATION_PROVIDERS = /*@ts2dart_const*/[
-  /* @ts2dart_Provider */ {provide: PlatformLocation, useClass: WebWorkerPlatformLocation},
+export const WORKER_APP_LOCATION_PROVIDERS = [
+  {provide: PlatformLocation, useClass: WebWorkerPlatformLocation},
   {
     provide: APP_INITIALIZER,
     useFactory: (platformLocation: WebWorkerPlatformLocation, zone: NgZone) => () =>
