+++ modules/@angular/platform-browser/test/xhr/xhr_impl_spec.ts
@@ -11,7 +11,6 @@ import {
 } from '@angular/core/testing/testing_internal';
 import {XHRImpl} from '../../src/xhr/xhr_impl';
 import {PromiseWrapper} from '../../src/facade/async';
-import {IS_DART} from '../../src/facade/lang';
 
 export function main() {
   describe('XHRImpl', () => {
@@ -23,8 +22,7 @@ export function main() {
     // will be relative to here, so url200 should look like
     // static_assets/200.html.
     // We currently have no way of detecting this.
-    var urlBase = IS_DART ? '' : '/base/modules/@angular/';
-    var url200 = urlBase + 'platform-browser/test/browser/static_assets/200.html';
+    var url200 = '/base/modules/@angular/platform-browser/test/browser/static_assets/200.html';
     var url404 = '/bad/path/404.html';
 
     beforeEach(() => { xhr = new XHRImpl(); });
