+++ modules/@angular/platform-browser/testing/benchmark_util.ts
@@ -1,7 +1,7 @@
 import {BrowserDomAdapter} from '../src/browser/browser_adapter';
 import {document, window} from '../src/facade/browser';
 import {NumberWrapper, isBlank} from '../src/facade/lang';
-import {BaseException, WrappedException} from '../src/facade/exceptions';
+import {BaseException} from '../src/facade/exceptions';
 
 var DOM = new BrowserDomAdapter();
